PR fixes

+++ internal/database/migrations/migrations.go
@@ -31,8 +31,8 @@ func Seed(db *gorm.DB) error {
 		Bio:           faker.Name().Title(),
 		OriginCountry: faker.Address().Country(),
 		CurrentCity:   "Toronto",
-		Latitude:      faker.Address().Latitude(),
-		Longitude:     faker.Address().Longitude(),
+		Latitude:      float64(faker.Address().Latitude()),
+		Longitude:     float64(faker.Address().Longitude()),
 		Stories: []models.Story{
 			{
 				Title:   faker.Hacker().Noun(),
